+++ packages/react/src/index.ts
@@ -17,6 +17,7 @@ export * from "./quickMenu";
 export * from "./questTracker";
 export * from "./notifications";
 export * from "./coachMarks";
+export * from "./timerReadout";
 export * from "./fastTravel";
 export * from "./components";
 export * from "./hud";